+++ lucene/pim-benchmarks/exact_phrase/src/SearchWikiMultiThread.java
@@ -69,7 +69,7 @@ public class SearchWikiMultiThread {
     String queries = null;
     int repeat = 0;
     String queryString = null;
-    int totalTime = 0;
+    long totalTime = 0;
     long cpuTime = 0;
 
     for (int i = 0; i < args.length; i++) {
